+++ alembic/README
@@ -1,9 +1,14 @@
+Инициализировать Alembic в проекте(должены использовать асинхронный шаблон)
+```bash
+poetry run alembic init --template async alembic 
+```
+
 Примените миграции
 ```bash
-alembic upgrade head
+poetry run alembic upgrade head
 ```
 
 Создание миграций, если внесены изменения в /models
 ```bash
-alembic revision --autogenerate -m "Your commit"
+poetry run alembic revision --autogenerate -m "Your commit"
 ```
