--- alembic/versions/182d5908e82a_update_user_model.py
@@ -1,91 +1,0 @@
-"""Update user model
-
-Revision ID: 182d5908e82a
-Revises: 
-Create Date: 2024-10-04 01:00:56.159813
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import postgresql
-
-# revision identifiers, used by Alembic.
-revision: str = '182d5908e82a'
-down_revision: Union[str, None] = None
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('checkcompanyportfolio',
-    sa.Column('project_name', sa.VARCHAR(length=48), nullable=False),
-    sa.Column('url', sa.VARCHAR(length=128), nullable=False),
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('contactmanager',
-    sa.Column('first_name', sa.VARCHAR(length=32), nullable=False),
-    sa.Column('phone_number', sa.VARCHAR(length=25), nullable=False),
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('info',
-    sa.Column('question_type', postgresql.ENUM('GENERAL_QUESTIONS', 'PROBLEMS_WITH_PRODUCTS', name='question_enum'), nullable=False),
-    sa.Column('question', sa.TEXT(), nullable=False),
-    sa.Column('answer', sa.TEXT(), nullable=False),
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('question')
-    )
-    op.create_table('informationaboutcompany',
-    sa.Column('name', sa.VARCHAR(length=48), nullable=False),
-    sa.Column('url', sa.VARCHAR(length=128), nullable=False),
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('productcategory',
-    sa.Column('title', sa.VARCHAR(length=150), nullable=False),
-    sa.Column('response', sa.TEXT(), nullable=False),
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('user',
-    sa.Column('tg_id', sa.BIGINT(), nullable=False),
-    sa.Column('name', sa.VARCHAR(length=32), nullable=True),
-    sa.Column('phone', sa.VARCHAR(length=25), nullable=True),
-    sa.Column('need_support', sa.BOOLEAN(), nullable=False),
-    sa.Column('need_contact_with_manager', sa.BOOLEAN(), nullable=False),
-    sa.Column('role', postgresql.ENUM('USER', 'ADMIN', 'MANAGER', name='role_enum'), nullable=False),
-    sa.Column('join_date', postgresql.TIMESTAMP(timezone=True), server_default=sa.text('now()'), nullable=False),
-    sa.Column('shipping_date', postgresql.TIMESTAMP(timezone=True), server_default=sa.text('now()'), nullable=False),
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('tg_id')
-    )
-    op.create_table('categorytype',
-    sa.Column('name', sa.VARCHAR(length=150), nullable=False),
-    sa.Column('product_id', sa.Integer(), nullable=False),
-    sa.Column('url', sa.VARCHAR(length=128), nullable=False),
-    sa.Column('media', sa.VARCHAR(length=128), nullable=True),
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.ForeignKeyConstraint(['product_id'], ['productcategory.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_categorytype_product_id'), 'categorytype', ['product_id'], unique=False)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f('ix_categorytype_product_id'), table_name='categorytype')
-    op.drop_table('categorytype')
-    op.drop_table('user')
-    op.drop_table('productcategory')
-    op.drop_table('informationaboutcompany')
-    op.drop_table('info')
-    op.drop_table('contactmanager')
-    op.drop_table('checkcompanyportfolio')
-    # ### end Alembic commands ###
